fix(TCOMP-1039): move classes related to Update feature to corresponding packages
