+++ main/plugins/org.talend.sdk.component.studio-integration/src/main/java/org/talend/sdk/component/studio/model/action/update/UpdateAction.java
@@ -13,7 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.talend.sdk.component.studio.model.action;
+package org.talend.sdk.component.studio.model.action.update;
+
+import org.talend.sdk.component.studio.model.action.Action;
 
 public class UpdateAction extends Action<Object> {
     public UpdateAction(final String actionName, final String family, final Type type) {
